move all files under src
