+++ server.py
@@ -9,46 +9,47 @@ from src.config.save_config_gradio import save_config_gradio
 from src.tortoise.generation_tab_tortoise import css_tortoise, generation_tab_tortoise
 from src.history_tab.history_css import history_css
 from src.load_config import default_config
-from settings_tab_gradio import settings_tab_gradio
+from src.settings_tab_gradio import settings_tab_gradio
 from src.bark_tab.generation_tab_bark import generation_tab_bark, bark_css
 from src.clone_tab.tab_voice_clone_demo import tab_voice_clone_demo
 import gradio as gr
 from src.history_tab.main import favorites_tab, history_tab, voices_tab
 from src.model_manager import model_manager
-from settings_tab_bark import settings_tab_bark
+from src.settings_tab_bark import settings_tab_bark
 from src.config.config import config
-from typing import Optional, Callable
 
 setup_or_recover.dummy()
 dotenv_init.init()
 
-def load_models(
-    text_use_gpu,
-    text_use_small,
-    coarse_use_gpu,
-    coarse_use_small,
-    fine_use_gpu,
-    fine_use_small,
-    codec_use_gpu
-):
 
+def load_models(
+        text_use_gpu,
+        text_use_small,
+        coarse_use_gpu,
+        coarse_use_small,
+        fine_use_gpu,
+        fine_use_small,
+        codec_use_gpu
+):
     save_config_bark(text_use_gpu,
-                text_use_small,
-                coarse_use_gpu,
-                coarse_use_small,
-                fine_use_gpu,
-                fine_use_small,
-                codec_use_gpu)
+                     text_use_small,
+                     coarse_use_gpu,
+                     coarse_use_small,
+                     fine_use_gpu,
+                     fine_use_small,
+                     codec_use_gpu)
     # download and load all models
-    # TODO: try catch for memmory errors
+    # TODO: try catch for memory errors
     model_manager.reload_models(config)
     return gr.Button.update(value="Reload models", interactive=True)
+
 
 full_css = ""
 full_css += material_symbols_css
 full_css += css_tortoise
 full_css += history_css
 full_css += bark_css
+
 
 def reload_config_and_restart_ui():
     os._exit(0)
@@ -59,7 +60,9 @@ def reload_config_and_restart_ui():
     # time.sleep(1)
     # demo.launch(**gradio_interface_options)
 
-gradio_interface_options = config["gradio_interface_options"] if "gradio_interface_options" in config else default_config
+
+gradio_interface_options = config[
+    "gradio_interface_options"] if "gradio_interface_options" in config else default_config
 
 with gr.Blocks(css=full_css) as demo:
     gr.Markdown("# TTS Generation WebUI (Bark & Tortoise)")
@@ -80,7 +83,8 @@ def print_pretty_options(options):
     print("Gradio interface options:")
     max_key_length = max(len(key) for key in options.keys())
     for key, value in options.items():
-        print(f"  {key}:{' '*(max_key_length - len(key))} {value}")
+        print(f"  {key}:{' ' * (max_key_length - len(key))} {value}")
+
 
 print("Starting Gradio server...")
 if not gradio_interface_options["enable_queue"]:
@@ -88,5 +92,5 @@ if not gradio_interface_options["enable_queue"]:
     gradio_interface_options["enable_queue"] = True
 print_pretty_options(gradio_interface_options)
 
-if __name__ == "__main__":   
+if __name__ == "__main__":
     demo.launch(**gradio_interface_options)
